refactor: extract filter logic to reusable utility
