+++ src/commands/visualizeProject.ts
@@ -1,9 +1,10 @@
 import * as vscode from 'vscode';
-import { createVisualizationPanel, generateMermaidDiagram } from '../visualization/visualizer';
+import { createVisualizationPanel } from '../visualization/visualizer';
 import { handleExport } from '../export/exportHandler';
 import { ContractGraph } from '../types/graph';
 import { detectLanguage, parseContractWithImports, getFunctionTypeFilters } from '../parser/parserUtils';
 import { logError } from '../logger';
+import { applyFilters } from './filterUtils';
 
 let panel: vscode.WebviewPanel | undefined;
 
@@ -59,75 +60,9 @@ export async function visualizeProject(context: vscode.ExtensionContext, fileUri
             panel!.webview.onDidReceiveMessage(
                 async (message) => {
                     if (message.command === 'applyFilters') {
-                        try {
-                            const selectedTypes = message.selectedTypes as string[];
-                            const nameFilter = message.nameFilter ? message.nameFilter.trim().toLowerCase() : '';
-
-                            const remainingNodeIds = new Set<string>();
-                            const nodesPassingTypeFilter = originalGraph.nodes.filter(node => {
-                                const nodeType = node.functionType || 'regular';
-                                const shouldKeep = selectedTypes.includes(nodeType);
-                                if (shouldKeep) {
-                                    remainingNodeIds.add(node.id);
-                                }
-                                return shouldKeep;
-                            });
-
-                            let filteredNodes = nodesPassingTypeFilter;
-                            if (nameFilter) {
-                                const matchingNodeIds = new Set<string>();
-                                const nameMatchedNodes = nodesPassingTypeFilter.filter(node => {
-                                    const label = node.label.toLowerCase();
-                                    const id = node.id.toLowerCase();
-
-                                    if (label.includes(nameFilter) || id.includes(nameFilter)) {
-                                        matchingNodeIds.add(node.id);
-                                        return true;
-                                    }
-                                    return false;
-                                });
-
-                                const connectedNodeIds = new Set<string>();
-                                originalGraph.edges.forEach(edge => {
-                                    if (matchingNodeIds.has(edge.from) && remainingNodeIds.has(edge.to)) {
-                                        connectedNodeIds.add(edge.to);
-                                    }
-                                    if (matchingNodeIds.has(edge.to) && remainingNodeIds.has(edge.from)) {
-                                        connectedNodeIds.add(edge.from);
-                                    }
-                                });
-
-                                const visibleNodeIds = new Set([...matchingNodeIds, ...connectedNodeIds]);
-                                filteredNodes = nodesPassingTypeFilter.filter(node =>
-                                    visibleNodeIds.has(node.id)
-                                );
-                            }
-
-                            const finalNodeIds = new Set(filteredNodes.map(node => node.id));
-                            const filteredEdges = originalGraph.edges.filter(edge =>
-                                finalNodeIds.has(edge.from) && finalNodeIds.has(edge.to)
-                            );
-
-                            const filteredGraph: ContractGraph = {
-                                nodes: filteredNodes,
-                                edges: filteredEdges,
-                            };
-
-                            const newMermaidDiagram = generateMermaidDiagram(filteredGraph);
-
-                            panel!.webview.postMessage({
-                                command: 'updateDiagram',
-                                diagram: newMermaidDiagram
-                            });
-
-                        } catch (filterError: any) {
-                            logError('Error applying filters', filterError);
-                            vscode.window.showErrorMessage(`Error applying filters: ${filterError.message || String(filterError)}`);
-                            panel!.webview.postMessage({
-                                command: 'filterError',
-                                error: filterError.message || String(filterError)
-                            });
-                        }
+                        const selectedTypes = message.selectedTypes as string[];
+                        const nameFilter = message.nameFilter ? message.nameFilter.trim().toLowerCase() : '';
+                        applyFilters(panel!, originalGraph, selectedTypes, nameFilter);
                     } else {
                         await handleExport(panel!, message, context);
                     }
